Sporadical wrapper for pybindings; Fixing eigen includes; Fixing namespaces
