+++ cpp/IK/IK_spherical_2_intersecting.cpp
@@ -12,7 +12,7 @@
 #include <iostream>
 #include "IK_spherical_2_intersecting.h"
 #include "../rand_cpp.h"
-#include "../subproblems/sp.cpp"
+#include "../subproblems/sp.h"
 
 namespace IKS {
 
@@ -24,7 +24,7 @@ namespace IKS {
 
     for (unsigned int i = 0; i < kin.joint_type.size(); i++) {
       if (kin.joint_type(i) == 0 || kin.joint_type(i) == 2) {
-        R = R * rot(kin.H.col(i), soln.Q[i][0]);
+        R = R * RAND_CPP::rot(kin.H.col(i), soln.Q[i][0]);
       } else if (kin.joint_type(i) == 1 || kin.joint_type(i) == 3) {
         p = p + R * kin.H.col(i)*soln.Q[i][0];
       }
@@ -40,21 +40,21 @@ namespace IKS {
     for (int i = 0; i < 6; i++) {
       soln.Q.push_back(std::vector<double>());
       for (int j = 0; j < 5; j++) {
-        soln.Q[i].push_back(rand_angle());
+        soln.Q[i].push_back(RAND_CPP::rand_angle());
       }
     }
 
 
     kin.joint_type = Eigen::Matrix<double, 1, 6>::Zero();
 
-    kin.H = rand_normal_vec(6);
-    kin.P.col(0) = rand_vec();
+    kin.H = RAND_CPP::rand_normal_vec(6);
+    kin.P.col(0) = RAND_CPP::rand_vec();
     kin.P.col(1) = zv;
-    kin.P.col(2) = rand_vec();
-    kin.P.col(3) = rand_vec();
+    kin.P.col(2) = RAND_CPP::rand_vec();
+    kin.P.col(3) = RAND_CPP::rand_vec();
     kin.P.col(4) = zv;
     kin.P.col(5) = zv;
-    kin.P.col(6) = rand_vec();
+    kin.P.col(6) = RAND_CPP::rand_vec();
 
 
     IKS::fwdkin(kin, soln, T, R);
@@ -69,28 +69,28 @@ namespace IKS {
     for (int i = 0; i < 6; i++) {
       soln.Q.push_back(std::vector<double>());
       for (int j = 0; j < 5; j++) {
-        soln.Q[i].push_back(rand_angle());
+        soln.Q[i].push_back(RAND_CPP::rand_angle());
       }
     }
 
     kin.joint_type = Eigen::Matrix<double, 1, 6>::Zero();
 
-    kin.H = rand_normal_vec(6);
-    kin.P.col(0) = rand_vec();
+    kin.H = RAND_CPP::rand_normal_vec(6);
+    kin.P.col(0) = RAND_CPP::rand_vec();
     kin.P.col(1) = zv;
-    kin.P.col(2) = rand_vec();
-    kin.P.col(3) = rand_vec();
+    kin.P.col(2) = RAND_CPP::rand_vec();
+    kin.P.col(3) = RAND_CPP::rand_vec();
     kin.P.col(4) = zv;
     kin.P.col(5) = zv;
     kin.P.col(6) = zv;
-    kin.H.col(1) = rand_perp_normal_vec(kin.H.col(1));
+    kin.H.col(1) = RAND_CPP::rand_perp_normal_vec(kin.H.col(1));
     kin.H.col(2) = kin.H.col(1);
     kin.P.col(3) = kin.P.col(3) - kin.H.col(2) * (kin.H.col(2).transpose() * (kin.P.col(2) + kin.P.col(3)));
-    kin.P.col(4) = rand_perp_normal_vec(kin.H.col(3));
-    kin.P.col(5) = rand_perp_normal_vec(kin.H.col(4));
+    kin.P.col(4) = RAND_CPP::rand_perp_normal_vec(kin.H.col(3));
+    kin.P.col(5) = RAND_CPP::rand_perp_normal_vec(kin.H.col(4));
 
-    R = rot(rand_normal_vec(), rand_angle());
-    T = rand_vec();
+    R = RAND_CPP::rot(RAND_CPP::rand_normal_vec(), RAND_CPP::rand_angle());
+    T = RAND_CPP::rand_vec();
 
   }
 
@@ -121,7 +121,7 @@ namespace IKS {
       t2.push_back(0);
 
       bool t12_is_ls = IKS::sp2_run(p_16,
-                              kin.P.col(2) + rot(kin.H.col(2), q3) * kin.P.col(3),
+                              kin.P.col(2) + RAND_CPP::rot(kin.H.col(2), q3) * kin.P.col(3),
                               -kin.H.col(0),
                               kin.H.col(1),
                               t1, t2);
@@ -131,7 +131,7 @@ namespace IKS {
         double q1 = t1[j];
         double q2 = t2[j];
 
-        Eigen::Matrix<double, 3, 3> R_36 = rot(-kin.H.col(2), q3) * rot(-kin.H.col(1), q2) * rot(-kin.H.col(0), q1) * R_0T;
+        Eigen::Matrix<double, 3, 3> R_36 = RAND_CPP::rot(-kin.H.col(2), q3) * RAND_CPP::rot(-kin.H.col(1), q2) * RAND_CPP::rot(-kin.H.col(0), q1) * R_0T;
 
         std::vector<double> t5;
 
@@ -147,16 +147,15 @@ namespace IKS {
 
           double q4, q6 = 0;
 
-          bool q4_is_ls = IKS::sp1_run(rot(kin.H.col(4), q5) * kin.H.col(5),
+          bool q4_is_ls = IKS::sp1_run(RAND_CPP::rot(kin.H.col(4), q5) * kin.H.col(5),
                                   R_36*kin.H.col(5),
                                   kin.H.col(4), q4);
 
 
-          bool q6_is_ls = IKS::sp1_run(rot(-kin.H.col(4), q5) * kin.H.col(3),
+          bool q6_is_ls = IKS::sp1_run(RAND_CPP::rot(-kin.H.col(4), q5) * kin.H.col(3),
                                   R_36.transpose()*kin.H.col(3),
                                   -kin.H.col(6), q6);
 
-          
           Eigen::Matrix<double, 6, 1> q;
           q << q1, q2, q3, q4, q5, q6;
           Q.conservativeResize(6, Q.cols() + 1);
@@ -165,9 +164,16 @@ namespace IKS {
           Eigen::Matrix<double, 5, 1> q_ls;
           q_ls << t3_is_ls, t12_is_ls, q5_is_ls, q4_is_ls, q6_is_ls;
           Q_LS.conservativeResize(5, Q_LS.cols() + 1);
-          Q_LS.col(Q_LS.cols() - 1) = q_ls;
+          Q_LS.col(Q_LS.cols() - 1) = q_ls; 
         }
       }
     }
+  }
+
+  Solution IK_spherical_2_intersecting_py(const Eigen::Matrix<double, 3, 3>& R_0T, const Eigen::Vector3d& p_0T, const Kin& kin)
+  {
+      Solution sol;
+      IK_spherical_2_intersecting(R_0T, p_0T, kin, sol.Q, sol.is_LS_vec);
+      return sol;
   }
 }
