+++ cpp/subproblems/sp.cpp
@@ -4,7 +4,11 @@
 // Purpose: Port of the subproblem files functionality
 //---------------------------------------------------------------//
 #include "sp.h"
-#include "../rand_cpp.h"
+#include <math.h>
+#include <vector>
+#include <complex>
+#include <eigen3/Eigen/Dense>
+#include <eigen3/unsupported/Eigen/Polynomials>
 
 namespace IKS {
 	void cone_polynomials(const Eigen::Vector3d &p0_i, const Eigen::Vector3d &k_i, const Eigen::Vector3d &p_i, const Eigen::Vector3d &p_i_s, const Eigen::Vector3d &k2,
